% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOG.R
\name{BOG}
\alias{BOG}
\title{calculate disease similarity by BOG}
\usage{
BOG(D1, D2, d2g)
}
\arguments{
\item{D1}{a vector consists disease ids}

\item{D2}{another vector consists disease ids}

\item{d2g}{a list of disease-gene associations}
}
\value{
a matrix of disease disease simialrity 
which rownames is D1 and colnames is D2
}
\description{
given two vectors of diseases and a list of disease-gene associations,
this function will calculate disease similarity by method BOG.
}
\examples{
data(d2g_separation) #get disease-gene associations
ds<-sample(names(d2g_separation),5)
sim<-BOG(ds,ds,d2g_separation)
Normalize(sim) #normalize BOG sim scores
}
\author{
Peng Ni, Min Li
}
\references{
Mathur S, Dinakarpandian D. Automated ontological gene annotation for computing disease similarity[J]. 
 AMIA Summits on Translational Science Proceedings, 2010, 2010: 12
}
\seealso{
\code{\link{Normalize}}
}

