% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dSimer-package.R
\docType{data}
\name{d2g_fundo_symbol}
\alias{d2g_fundo_symbol}
\title{disease-gene associations}
\description{
a list of disease-gene associations from FunDO which diseases are DOIDs and
genes are gene symbols
}
\references{
Osborne J D, Flatow J, Holko M, et al. Annotating the human genome 
with Disease Ontology[J]. BMC genomics, 2009, 10(Suppl 1): S6.
}
\keyword{dataset}

