% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dSimer-package.R
\docType{data}
\name{d2g_separation}
\alias{d2g_separation}
\title{disease-gene associations}
\description{
a list of disease-gene associations from the reference paper (see below) 
which diseases are disease names and genes are entrez ids
}
\references{
Menche J, Sharma A, Kitsak M, et al. Uncovering disease-disease 
relationships through the incomplete interactome[J]. Science, 2015, 
347(6224): 1257601.
}
\keyword{dataset}

