% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dSimer-package.R
\docType{data}
\name{d2go_sample}
\alias{d2go_sample}
\title{a sample of disease-GO term associations}
\description{
a sample list of disease-GO term associations, the entire data of disease-
GO term associations can be obtained by function HypergeometricTest
}
\seealso{
\code{\link{HypergeometricTest}}
}
\keyword{dataset}

