% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dSimer-package.R
\docType{data}
\name{d2s_hsdn}
\alias{d2s_hsdn}
\title{d2s_hsdn}
\value{
d2s_hsdn is a data.frame of 73726 rows and 3 columns, contains PubMed co-
occurrences of diseases and symptoms, will be used in method CosineDFV.
}
\description{
diseases, symptoms and their co-occurrences in PubMed
}
\examples{
data(d2s_hsdn)
}
\references{
Zhou X Z, Menche J, Barabasi A L, et al. Human symptoms-disease 
network[J]. Nature communications, 2014, 5.
}
\seealso{
\code{\link{CosineDFV}}
}
\keyword{dataset}

