% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dSimer-package.R
\docType{data}
\name{d2s_hsdn_sample}
\alias{d2s_hsdn_sample}
\title{d2s_hsdn_sample}
\value{
d2s_hsdn__sample is a data.frame of 1480 rows and 3 columns, contains PubMed co-
occurrences of diseases and symptoms. It is a sample of d2s_hsdn.
}
\description{
a sample of d2s_hsdn
}
\examples{
data(d2s_hsdn_sample)
}
\references{
Zhou X Z, Menche J, Barabasi A L, et al. Human symptoms-disease 
network[J]. Nature communications, 2014, 5.
}
\seealso{
\code{\link{d2s_hsdn}}, \code{\link{CosineDFV}}
}
\keyword{dataset}

