% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SemFunSim.R
\name{FunSim}
\alias{FunSim}
\title{calculate disease similarity by FunSim}
\usage{
FunSim(D1, D2, d2g, LLSnList)
}
\arguments{
\item{D1}{a vector consists disease ids}

\item{D2}{another vector consists disease ids}

\item{d2g}{a list of disease-gene associations, while gene ids should be entrez id.}

\item{LLSnList}{a list of gene-gene log-likelihood score from HumanNet}
}
\value{
a matrix of disease disease simialrity
which rownames is D1 and colnames is D2
}
\description{
given two vectors of diseases, a list of disease-gene associations ,
and a list of gene-gene log-likelihood score from HumanNet, 
this function will calculate disease similarity by method FunSim
}
\examples{
## in this method, we must use disease-gene associations 
## which genes are represented by entrez ids because of
## HumanNet
data(d2g_fundo_entrezid)
data(HumanNet_sample)
## we specified 5 DOIDs to match Human_sample
ds<-c("DOID:8176","DOID:2394","DOID:3744","DOID:8466","DOID:5679")
llsnlist<-LLSn2List(HumanNet_sample)
FunSim(ds,ds,d2g_fundo_entrezid,llsnlist)
}
\author{
Peng Ni, Min Li
}
\references{
Cheng L, Li J, Ju P, et al. SemFunSim: a new method for measuring 
disease similarity by integrating semantic and gene functional association[J]. 
PloS one, 2014, 9(6): e99415.
}
\seealso{
\code{\link{LLSn2List}}
}

