% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dSimer-package.R
\docType{data}
\name{go2g_sample}
\alias{go2g_sample}
\title{a sample of GO term-gene associations}
\description{
a sample list of GO term-gene associations, the entire data of GO term-gene
can be obtained by function get_GOterm2GeneAssos
}
\seealso{
\code{\link{get_GOterm2GeneAssos}}
}
\keyword{dataset}

