% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dSimer-package.R
\docType{data}
\name{graphlet_sig_hprd}
\alias{graphlet_sig_hprd}
\title{graphlet signature}
\description{
matrix, graphlet signatures of nodes in HPRD PPI network calculated by 
ORCA tool, will be using in method Sun_topology
}
\references{
Hocevar T, Demsar J. A combinatorial approach to graphlet 
counting[J]. Bioinformatics, 2014, 30(4): 559-565.
}
\seealso{
\code{\link{Sun_topology}}
}
\keyword{dataset}

