% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dSimer-package.R
\docType{data}
\name{HumanNet_sample}
\alias{HumanNet_sample}
\title{a sample of HumanNet likelihood score data, will be used in method FunSim}
\description{
data.frame, where each row indicates a pair of genes and 
their normalized likelihood score in HumanNet, will be used in
method FunSim after being converted to list by method LLSn2List
}
\references{
Cheng L, Li J, Ju P, et al. SemFunSim: a new method for measuring 
disease similarity by integrating semantic and gene functional association[J]. 
PloS one, 2014, 9(6): e99415.
}
\seealso{
\code{\link{FunSim}}, \code{\link{LLSn2List}}
}
\keyword{dataset}

