% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICod.R
\name{ICod}
\alias{ICod}
\title{calculate disease similarity by ICod}
\usage{
ICod(D1, D2, d2g, graph, A = 0.9, b = 1, C = 0)
}
\arguments{
\item{D1}{a vector consists disease ids}

\item{D2}{another vector consists disease ids}

\item{d2g}{a list of disease-gene associations}

\item{graph}{an igraph graph object of PPI network}

\item{A}{a parameter used in ICod to calculate transformed distance of node pair, default 0.9}

\item{b}{a parameter used in ICod to calculate transformed distance of node pair, default 1}

\item{C}{a parameter used in ICod to calculate disease similarity, default 0}
}
\value{
a matrix of disease disease simialrity
which rownames is D1 and colnames is D2
}
\description{
given two vectors of diseases, a list of disease-gene associations 
and a PPI neteork, this function will calculate disease similarity by method ICod
}
\examples{
data(d2g_fundo_symbol)
data(PPI_HPRD)

graph_hprd<-graph.data.frame(PPI_HPRD,directed=FALSE) #get a igraph object based on HPRD data
ds<-sample(names(d2g_fundo_symbol),5)
ICod(ds,ds,d2g_fundo_symbol,graph_hprd)
}
\author{
Peng Ni, Min Li
}
\references{
Paik H, Heo HS, Ban H, et al. Unraveling human protein interaction 
networks underlying co-occurrences of diseases and pathological conditions[J]. 
Journal of translational medicine, 2014, 12(1): 99.
}

