% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_use.R
\name{InformationContent}
\alias{InformationContent}
\title{calculating information content}
\usage{
InformationContent(T2G)
}
\arguments{
\item{T2G}{a list of Term-Gene associations which names are term ids}
}
\value{
a list of IC values of inputted term ids
}
\description{
calculate information content of all term ids in a term list
}
\examples{
data(d2g_fundo_symbol)
InformationContent(d2g_fundo_symbol[1:5])
}
\author{
Peng Ni
}

