% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dSimer-package.R
\docType{data}
\name{interactome}
\alias{interactome}
\title{interactome data}
\description{
a data.frame, contains interactome data in reference below
}
\references{
Menche J, Sharma A, Kitsak M, et al. Uncovering disease-disease 
relationships through the incomplete interactome[J]. Science, 2015, 
347(6224): 1257601.
}
\keyword{dataset}

