% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_use.R
\name{jaccardindex}
\alias{jaccardindex}
\title{calculating Jaccard Index}
\usage{
jaccardindex(x1, x2, x2y)
}
\arguments{
\item{x1}{a disease id}

\item{x2}{another disease id}

\item{x2y}{a list of disease-gene associations which consists x1 and x2}
}
\value{
numeric value of a jaccard index of x1 and x2
}
\description{
calculate Jaccard Index of two terms by using their annotated genes
}
\examples{
## this function is not just for disease-gene associations
data(d2go_sample)
d1<-names(d2go_sample)[1]
d2<-names(d2go_sample)[2]
jaccardindex(d1,d2,d2go_sample)
}
\author{
Peng Ni
}

