% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_use.R
\name{Normalize}
\alias{Normalize}
\title{normalize data}
\usage{
Normalize(data)
}
\arguments{
\item{data}{a numeric/integer vector or matrix}
}
\value{
normalized vector or matrix
}
\description{
normalize a vector or a matrix based on the formula from SemFunSim
}
\examples{
sim<-matrix(1:9,3,3)
Normalize(sim)
}
\author{
Peng Ni
}
\references{
Cheng L, Li J, Ju P, et al. SemFunSim: a new method for measuring 
disease similarity by integrating semantic and gene functional association[J]. 
PloS one, 2014, 9(6): e99415.
}

