% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dSimer-package.R
\docType{data}
\name{orbit_dependency_count}
\alias{orbit_dependency_count}
\title{orbit_dependency_count}
\value{
orbit_dependency_count is a 73-dim vector, indicating 73 orbits' 
dependency count in graphlet theory, used to calculate weight factor in 
method setWeight.
}
\description{
orbit dependency count
}
\examples{
data(orbit_dependency_count)
}
\references{
Milenkovic T, Przulj N. Uncovering biological network function via 
graphlet degree signatures[J]. Cancer informatics, 2008, 6: 257.
}
\seealso{
\code{\link{setWeight}}
}
\keyword{dataset}

