% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{similarity matrix heatmap plotting}
\usage{
plot_heatmap(simmat, xlab = "", ylab = "", color.low = "white",
  color.high = "red", labs = TRUE, digits = 2, labs.size = 3,
  font.size = 14)
}
\arguments{
\item{simmat}{a similarity matrix}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{color.low}{color of low value}

\item{color.high}{color of high value}

\item{labs}{logical, add text label or not}

\item{digits}{round digit numbers}

\item{labs.size}{lable size}

\item{font.size}{font size}
}
\value{
a ggplot object
}
\description{
plot heatmap of a disease similarity matrix
}
\examples{

data(d2g_separation)
data(interactome)

graph_interactome<-graph.data.frame(interactome,directed=FALSE)
ds<-c("myocardial ischemia","myocardial infarction","coronary artery disease",
 "cerebrovascular disorders","arthritis, rheumatoid","diabetes mellitus, type 1",
 "autoimmune diseases of the nervous system","demyelinating autoimmune diseases, cns",
 "respiratory hypersensitivity","asthma","retinitis pigmentosa",
 "retinal degeneration","macular degeneration")
 
sep<-Separation(ds,ds,d2g_separation,graph_interactome)
sim<-Separation2Similarity(sep)
plot_heatmap(sim)
}
\author{
Peng Ni
}
\references{
Yu G, Wang L G, Yan G R, et al. DOSE: an R/Bioconductor package for 
disease ontology semantic and enrichment analysis[J]. 
Bioinformatics, 2015, 31(4): 608-609.
}

