% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot_net}
\alias{plot_net}
\title{plot a network based on a symmetric disease similarity matrix}
\usage{
plot_net(simmat, cutoff = 1, vertex.label.font = 2,
  vertex.label.dist = 0.5, vertex.label.color = "black",
  vertex.label.cex = 0.8, vertex.shape = "circle", vertex.size = 1,
  vertexsizefold = 5, vertex.color = "paleturquoise", edge.color = "red",
  layout = layout.fruchterman.reingold)
}
\arguments{
\item{simmat}{a symmetric similarity matrix}

\item{cutoff}{a cutoff value, only disease pairs have similarity scores 
no less than cutoff will be visualized in the network}

\item{vertex.label.font}{font size}

\item{vertex.label.dist}{font dist}

\item{vertex.label.color}{font text color}

\item{vertex.label.cex}{cex of vertex label}

\item{vertex.shape}{vertex shape}

\item{vertex.size}{vertex size}

\item{vertexsizefold}{a fold of vertex size}

\item{vertex.color}{vertex color}

\item{edge.color}{edge color}

\item{layout}{layout}
}
\value{
a igraph plot object
}
\description{
plot a network/graph of a symmetric disease similarity matrix, note that a unsymmetric
matrix can't be visualized into a network by this method.
}
\examples{

data(d2g_separation)
data(interactome)

graph_interactome<-graph.data.frame(interactome,directed=FALSE)
ds<-c("myocardial ischemia","myocardial infarction","coronary artery disease",
 "cerebrovascular disorders","arthritis, rheumatoid","diabetes mellitus, type 1",
 "autoimmune diseases of the nervous system","demyelinating autoimmune diseases, cns",
 "respiratory hypersensitivity","asthma","retinitis pigmentosa",
 "retinal degeneration","macular degeneration")
 
sep<-Separation(ds,ds,d2g_separation,graph_interactome)
sim<-Separation2Similarity(sep)
plot_net(sim,cutoff=0.2)
}
\author{
Peng Ni
}

