% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dSimer-package.R
\docType{data}
\name{PPI_HPRD}
\alias{PPI_HPRD}
\title{PPI_HPRD}
\value{
PPI_HPRD is a data.frame of 36867 rows and 2 columns. Each rows indicates
an interaction of two gene symbols. It was fetched from HPRD.
}
\description{
PPI data from HPRD
}
\examples{
data(PPI_HPRD)
}
\references{
Prasad T S K, Goel R, Kandasamy K, et al. Human protein reference 
database-2009 update[J]. Nucleic acids research, 2009, 37(suppl 1): D767-D772.
}
\keyword{dataset}

