% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSB.R
\name{PSB}
\alias{PSB}
\title{calculate disease similarity by PSB}
\usage{
PSB(D1, D2, d2go, go2g)
}
\arguments{
\item{D1}{a vector consists disease ids}

\item{D2}{another vector consists disease ids}

\item{d2go}{a list of disease-go associations}

\item{go2g}{a list of go-gene associations}
}
\value{
a matrix of disease disease simialrity 
which rownames is D1 and colnames is D2
}
\description{
given two vectors of diseases, a list of disease-GO term associations and a 
list of GO term-gene associations, this function will calculate disease 
similarity by method PSB
}
\examples{
## these are samples of GO-gene associations and disease-GO associations
data(go2g_sample)
data(d2go_sample)

##### the entire associations can be obtained by follows:
## go2g<-get_GOterm2GeneAssos(GOONTOLOGY = "BP") #get go-gene associations
## ## in this method, we must use disease-gene associations
## ## which genes are represented by gene symbols
## data(d2g_fundo_symbol)
## d2go<-HypergeometricTest(d2g = d2g_fundo_symbol,go2g = go2g)
##### ###################################################

ds<-names(d2go_sample)
sim<-PSB(ds,ds,d2go_sample,go2g_sample)
Normalize(sim)
}
\author{
Peng Ni, Min Li
}
\references{
Mathur S, Dinakarpandian D. Finding disease similarity based on implicit semantic similarity[J]. 
Journal of biomedical informatics, 2012, 45(2): 363-371.
}
\seealso{
\code{\link{get_GOterm2GeneAssos}}, \code{\link{HypergeometricTest}}, \code{\link{Normalize}}
}

