% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Separation.R
\name{Separation}
\alias{Separation}
\title{calculating network-based separation of disease pairs}
\usage{
Separation(D1, D2, d2g, graph)
}
\arguments{
\item{D1}{a vector consists disease ids}

\item{D2}{another vector consists disease ids}

\item{d2g}{a list of disease-gene associations}

\item{graph}{an igraph graph object of PPI network}
}
\value{
a matrix of disease disease network-based separation
which rownames is D1 and colnames is D2
}
\description{
given two vectors of diseases, a list of disease-gene associations
and a PPI network, this function will calculate network-based separation
by method Separation.
}
\examples{
data(d2g_separation)
data(interactome)

graph_interactome<-graph.data.frame(interactome,directed=FALSE)
ds<-sample(names(d2g_separation),5)
sep<-Separation(ds,ds,d2g_separation,graph_interactome)
sim<-Separation2Similarity(sep)
sim
}
\author{
Peng Ni, Min Li
}
\references{
Menche J, Sharma A, Kitsak M, et al. Uncovering disease-disease 
relationships through the incomplete interactome[J]. Science, 2015, 
347(6224): 1257601.
}
\seealso{
\code{\link{Separation2Similarity}}
}

