% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Separation.R
\name{Separation2Similarity}
\alias{Separation2Similarity}
\title{a method which convert separation to similarity}
\usage{
Separation2Similarity(data)
}
\arguments{
\item{data}{a numeric/integer matrix calculated by method Separation}
}
\value{
a similarity matrix
}
\description{
convert a separation matrix to a similarity matrix
}
\examples{
## see examples in function Separation
}
\author{
Peng Ni
}
\seealso{
\code{\link{Separation}}
}

