% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sun.R
\name{setWeight}
\alias{setWeight}
\title{set weight factor}
\usage{
setWeight(orbit_dependency_count)
}
\arguments{
\item{orbit_dependency_count}{a vector which each element are the dependency count of each orbit}
}
\value{
a vector which contains weight factors to each orbit
}
\description{
set weight factor of 73-orbits in graphlet theory
}
\examples{
data(orbit_dependency_count)
setWeight(orbit_dependency_count)
}
\author{
Peng Ni
}
\references{
Milenkovic T, Przulj N. Uncovering biological network function via 
graphlet degree signatures[J]. Cancer informatics, 2008, 6: 257.
}

