% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sun.R
\name{Sun_annotation}
\alias{Sun_annotation}
\title{Sun's annotation measure of disease similarity calculating}
\usage{
Sun_annotation(D1, D2, d2g)
}
\arguments{
\item{D1}{a vector consists disease ids}

\item{D2}{another vector consists disease ids}

\item{d2g}{a list of disease-gene associations}
}
\value{
a matrix of disease disease simialrity
which rownames is D1 and colnames is D2
}
\description{
given two vectors of diseases and a list of disease-gene associations,
this function will calculate disease similarity by method Sun_annotation
}
\examples{
data(d2g_separation)
ds<-sample(names(d2g_separation),5)
Sun_annotation(ds,ds,d2g_separation)
}
\author{
Peng Ni, Min Li
}
\references{
Sun K, Goncalves JP, Larminie C. Predicting disease associations 
via biological network analysis[J]. BMC bioinformatics, 2014, 15(1): 304.
}

