% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sun.R
\name{Sun_topology}
\alias{Sun_topology}
\title{Sun's topology measure of disease similarity calculating}
\usage{
Sun_topology(D1, D2, d2g, graphlet_sig_mat, weight)
}
\arguments{
\item{D1}{a vector consists disease ids}

\item{D2}{another vector consists disease ids}

\item{d2g}{a list of disease-gene associations}

\item{graphlet_sig_mat}{matrix of graphlet signature of nodes in a ppi network calculated by orca, see examples below.}

\item{weight}{a vector which elements are weight factors to each orbit in graphlet theory}
}
\value{
a disease disease similarity matrix
}
\description{
given two vectors of diseases, a list of disease-gene associations,
a matrix of genes' graphlet signature in a PPI network and a weight vector of
73 orbits in graphlet theory, this function will calculate disease similarity by 
method Sun_function
}
\examples{
data(d2g_fundo_symbol)
data(graphlet_sig_hprd) #get graphlet signatures of genes in HPRD PPI network
data(weight)
ds<-sample(names(d2g_fundo_symbol),5)
Sun_topology(ds,ds,d2g_fundo_symbol,graphlet_sig_hprd,weight)
}
\author{
Peng Ni, Min Li
}
\references{
Sun K, Goncalves JP, Larminie C. Predicting disease associations 
via biological network analysis[J]. BMC bioinformatics, 2014, 15(1): 304.
}

