% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dSimer-package.R
\docType{data}
\name{weight}
\alias{weight}
\title{weight factor}
\description{
a 73-dim vector, indicating 73 orbits' weight factor, will be used in
method Sun_topology
}
\references{
Sun K, Goncalves JP, Larminie C. Predicting disease associations 
via biological network analysis[J]. BMC bioinformatics, 2014, 15(1): 304.
}
\seealso{
\code{\link{setWeight}}, \code{\link{Sun_topology}}
}
\keyword{dataset}

