% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{x2y_conv2_y2x}
\alias{x2y_conv2_y2x}
\title{convert x2ylist to y2xlist}
\usage{
x2y_conv2_y2x(x2ylist)
}
\arguments{
\item{x2ylist}{a list which the names are xs and the elements are ys of each x}
}
\value{
a list of y2x
}
\description{
convert list of x-y associations to list of y-x associations
}
\examples{
data(go2g_sample)
g2go_sample<-x2y_conv2_y2x(go2g_sample[1:100])
}
\author{
Peng Ni
}

