% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dSimer-package.R
\docType{data}
\name{d2g_fundo_symbol}
\alias{d2g_fundo_symbol}
\title{d2g_fundo_symbol}
\value{
d2g_fundo_symbol is a named list of length 1855 which stored disease-gene 
associations from FunDO. The names are the DOIDs (DOIDs are ids of terms 
in Disease Ontology, e.g. "DOID:4" ) and list elements are vectors of gene symbols.
}
\description{
a list of disease-gene associations from FunDO.
}
\examples{
data(d2g_fundo_symbol)
}
\references{
Osborne J D, Flatow J, Holko M, et al. Annotating the human genome 
with Disease Ontology[J]. BMC genomics, 2009, 10(Suppl 1): S6.
}
\keyword{dataset}

