% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dSimer-package.R
\docType{data}
\name{d2g_separation}
\alias{d2g_separation}
\title{d2g_separation}
\value{
d2g_separation is a named list of length 299 which stored disease-gene
associations from the reference paper (see below). The names are diseases and 
list elements are vectors of gene entrez ids.
}
\description{
a list of disease-gene associations from the reference paper (see below).
}
\examples{
data(d2g_separation)
}
\references{
Menche J, Sharma A, Kitsak M, et al. Uncovering disease-disease 
relationships through the incomplete interactome[J]. Science, 2015, 
347(6224): 1257601.
}
\keyword{dataset}

