% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dSimer-package.R
\docType{data}
\name{d2go_sample}
\alias{d2go_sample}
\title{d2go_sample}
\value{
d2go_sample is a named list of length 3. The names are are the DOIDs 
(DOIDs are ids of terms in Disease Ontology, e.g. "DOID:4" ) and list 
elements are vectors of GO term ids.
The entire data of disease-GO term associations can be obtained by function HypergeometricTest.
}
\description{
a sample list of disease-GO term associations.
}
\examples{
data(d2go_sample)
}
\seealso{
\code{\link{HypergeometricTest}}
}
\keyword{dataset}

