% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dSimer-package.R
\docType{package}
\name{dSimer-package}
\alias{dSimer}
\alias{dSimer-package}
\title{Integration of Disease Similarity Methods}
\description{
dSimer is an R package which provides eight function-based methods 
for disease similarity calculation. These eight methods take advantage of 
diverse biological data which calculate disease similarity from different 
perspectives. The disease similarity matrix obtained from these eight methods 
can also be visualized by dSimer.
}
\details{
\tabular{ll}{ Package: \tab dSimer\cr Type: \tab Package\cr Version: \tab
0.99.2\cr Date: \tab 12-10-2015\cr biocViews:\tab Bioinformatics,
Similarity\cr Depends: \tab R (>= 3.3.0), igraph (>= 1.0.1)\cr Imports: \tab stats, Rcpp (>= 0.11.3), 
ggplot2, reshape2, GO.db, AnnotationDbi, org.Hs.eg.db\cr 
Suggests: \tab knitr, rmarkdown, BiocStyle\cr LinkingTo: \tab Rcpp\cr License: \tab
GPL (>= 2)\cr }
}
\author{
Peng Ni, Min Li
}
\keyword{package}

