% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{get_GOterm2GeneAssos}
\alias{get_GOterm2GeneAssos}
\title{get GO-gene associations}
\usage{
get_GOterm2GeneAssos(GOONTOLOGY = c("BP", "MF", "CC"), rm.IEAs = TRUE,
  rm.termlessthan3genes = TRUE)
}
\arguments{
\item{GOONTOLOGY}{"BP" or "MF" or "CC}

\item{rm.IEAs}{logical value, remove GO terms with evidence "IEA" or not}

\item{rm.termlessthan3genes}{logical value, remove terms whose number of annotated genes are less than 3 or not}
}
\value{
a list which names are GO term IDs and elements are gene symbols annotated with GO terms
}
\description{
get GO-gene associations from GO.db and org.Hs.eg.db
}
\examples{
go2g<-get_GOterm2GeneAssos(GOONTOLOGY="BP")
}
\author{
Peng Ni
}
\references{
Mathur S, Dinakarpandian D. Finding disease similarity based on implicit semantic similarity[J]. 
Journal of biomedical informatics, 2012, 45(2): 363-371.
}
\seealso{
\code{\link{PSB}}, \code{\link{Sun_function}}
}

