% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dSimer-package.R
\docType{data}
\name{go2g_sample}
\alias{go2g_sample}
\title{go2g_sample}
\value{
go2g_sample is a named list of length 465. The names are GO term ids (GOIDs)
and list elements are vectors of gene symbols. 
The entire data of GO term-gene assos can be obtained by function get_GOterm2GeneAssos.
}
\description{
a sample list of GO term-gene associations.
}
\examples{
data(go2g_sample)
}
\seealso{
\code{\link{get_GOterm2GeneAssos}}
}
\keyword{dataset}

