% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dSimer-package.R
\docType{data}
\name{interactome}
\alias{interactome}
\title{interactome}
\value{
interactome is a data.frame of 141296 rows and 2 columns. Each row indicates
an interaction of two gene entrez ids.
It was obtained from the reference below.
}
\description{
interactome data
}
\examples{
data(interactome)
}
\references{
Menche J, Sharma A, Kitsak M, et al. Uncovering disease-disease 
relationships through the incomplete interactome[J]. Science, 2015, 
347(6224): 1257601.
}
\keyword{dataset}

